refact class Filter and enum FilterType

+++ FilesFilter/src/ru/cft/drozdetskiy/FilterType.java
@@ -2,5 +2,5 @@ package ru.cft.drozdetskiy;
 
 public enum FilterType {
 
-    INTEGER, DOUBLE, STRING
+    LONG, DOUBLE, STRING
 }
